Refactor package structure: update imports to reflect new module organization and rename files for consistency
