+++ src/wembed_core/controllers/indexing_result_controller.py
@@ -5,7 +5,7 @@ wembed_core/controllers/scan_result_controller.py
 from typing import List, Optional
 
 from wembed_core.database import DatabaseService
-from wembed_core.models.indexing_results import FileIndexingResults
+from wembed_core.models.indexing.indexing_results import FileIndexingResults
 from wembed_core.schemas.indexing_result_schemas import IndexingResultSchema
