+++ src/wembed_core/models/__init__.py
@@ -1,6 +1,11 @@
+from wembed_core.models.indexing.indexed_files import IndexedFiles  # noqa: F401
+from wembed_core.models.indexing.indexed_obsidian_vaults import (  # noqa: F401
+    IndexedObsidianVaults,
+)
+from wembed_core.models.indexing.indexed_repos import IndexedRepos  # noqa: F401
+from wembed_core.models.indexing.indexing_results import (  # noqa: F401
+    FileIndexingResults,
+)
+
 from .code_chunker import *  # noqa: F403, F401
 from .dl_doc import *  # noqa: F403, F401
-from .indexed_files import IndexedFiles  # noqa: F401
-from .indexed_obsidian_vaults import IndexedObsidianVaults  # noqa: F401
-from .indexed_repos import IndexedRepos  # noqa: F401
-from .indexing_results import FileIndexingResults  # noqa: F401
