+++ tests/test_db_models/test_indexing_results_model.py
@@ -8,8 +8,8 @@ from datetime import datetime
 import pytest
 
 from wembed_core.config import AppConfig
-from wembed_core.database import AppBase, DatabaseService
-from wembed_core.models.indexing_results import FileIndexingResults
+from wembed_core.database import DatabaseService
+from wembed_core.models.indexing.indexing_results import FileIndexingResults
 
 
 class TestScanResultRecord:
